// src/openapi/model/city.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Country } from './country';


export interface City { 
    country?: Country;
    createdDate?: string;
    id?: number;
    lastModifiedBy?: string;
    lastModifiedDate?: string;
    name?: string;
    recordStatus?: number;
    zipCode?: string;
}